split model-headers file
