+++ app/services/matchers/giovanni/model3.js
@@ -1,6 +1,6 @@
 const logger = require("../../../utilities/logger");
 const matcherResult = require("../../matcher-result");
-const headers = require("../../model-headers");
+const headers = require("../../model-headers-pdf");
 const regex = require("../../../utilities/regex");
 const path = require("path");
 const filenameForLogging = path.join("app", __filename.split("app")[1]);
