+++ app/services/parsers/mands/model1.js
@@ -2,7 +2,7 @@ const parserModel = require("../../parser-model");
 const combineParser = require("../../parser-combine");
 const { mapPdfParser } = require("../../parser-map");
 const logger = require("../../../utilities/logger");
-const headers = require("../../model-headers");
+const headers = require("../../model-headers-pdf");
 const regex = require("../../../utilities/regex");
 const path = require("path");
 const filenameForLogging = path.join("app", __filename.split("app")[1]);
